Refactor out fill_order_json
