+++ ts/components/fill_order.tsx
@@ -30,6 +30,7 @@ import {Identicon} from 'ts/components/ui/identicon';
 import {EthereumAddress} from 'ts/components/ui/ethereum_address';
 import {TokenAmountInput} from 'ts/components/inputs/token_amount_input';
 import {FillWarningDialog} from 'ts/components/fill_warning_dialog';
+import {FillOrderJSON} from 'ts/components/fill_order_json';
 import {VisualOrder} from 'ts/components/visual_order';
 import {SchemaValidator} from 'ts/schemas/validator';
 import {orderSchema} from 'ts/schemas/order_schema';
@@ -113,7 +114,13 @@ export class FillOrder extends React.Component<FillOrderProps, FillOrderState> {
                                 Paste an order JSON snippet below to begin
                             </div>
                             <div className="pb2">Order JSON</div>
-                            {this.renderOrderJson()}
+                            <FillOrderJSON
+                                blockchain={this.props.blockchain}
+                                tokenByAddress={this.props.tokenByAddress}
+                                networkId={this.props.networkId}
+                                orderJSON={this.state.orderJSON}
+                                onFillOrderJSONChanged={this.onFillOrderJSONChanged.bind(this)}
+                            />
                             {this.renderOrderJsonNotices()}
                         </div>
                     }
@@ -132,7 +139,13 @@ export class FillOrder extends React.Component<FillOrderProps, FillOrderState> {
                                     showExpandableButton={true}
                                 />
                                 <CardText expandable={true}>
-                                    {this.renderOrderJson()}
+                                    <FillOrderJSON
+                                        blockchain={this.props.blockchain}
+                                        tokenByAddress={this.props.tokenByAddress}
+                                        networkId={this.props.networkId}
+                                        orderJSON={this.state.orderJSON}
+                                        onFillOrderJSONChanged={this.onFillOrderJSONChanged.bind(this)}
+                                    />
                                 </CardText>
                             </Card>
                             {this.renderOrderJsonNotices()}
@@ -170,52 +183,6 @@ export class FillOrder extends React.Component<FillOrderProps, FillOrderState> {
                 {!_.isEmpty(this.state.orderJSONErrMsg) &&
                     <Alert type={AlertTypes.ERROR} message={this.state.orderJSONErrMsg} />
                 }
-            </div>
-        );
-    }
-    private renderOrderJson() {
-        const addresses = _.keys(this.props.tokenByAddress);
-        const exchangeContract = this.props.blockchain.getExchangeContractAddressIfExists();
-        const hintSideToAssetToken = {
-            [Side.deposit]: {
-                amount: new BigNumber(35),
-                address: addresses[0],
-            },
-            [Side.receive]: {
-                amount: new BigNumber(89),
-                address: addresses[1],
-            },
-        };
-        const hintOrderExpiryTimestamp = utils.initialOrderExpiryUnixTimestampSec();
-        const hintSignatureData = {
-            hash: '0xf965a9978a0381ab58f5a2408ad967c...',
-            r: '0xf01103f759e2289a28593eaf22e5820032...',
-            s: '937862111edcba395f8a9e0cc1b2c5e12320...',
-            v: 27,
-        };
-        const hintSalt = ZeroEx.generatePseudoRandomSalt();
-        const hintOrder = utils.generateOrder(this.props.networkId, exchangeContract, hintSideToAssetToken,
-                                              hintOrderExpiryTimestamp, '', '', constants.MAKER_FEE,
-                                              constants.TAKER_FEE, constants.FEE_RECIPIENT_ADDRESS,
-                                              hintSignatureData, this.props.tokenByAddress, hintSalt);
-        const hintOrderJSON = `${JSON.stringify(hintOrder, null, '\t').substring(0, 500)}...`;
-        return (
-            <div>
-                <Paper className="p1 overflow-hidden" style={{height: 164}}>
-                    <TextField
-                        id="orderJSON"
-                        hintStyle={{bottom: 0, top: 0}}
-                        fullWidth={true}
-                        value={this.state.orderJSON}
-                        onChange={this.onFillOrderChanged.bind(this)}
-                        hintText={hintOrderJSON}
-                        multiLine={true}
-                        rows={6}
-                        rowsMax={6}
-                        underlineStyle={{display: 'none'}}
-                        textareaStyle={{marginTop: 0}}
-                    />
-                </Paper>
             </div>
         );
     }
@@ -293,7 +260,7 @@ export class FillOrder extends React.Component<FillOrderProps, FillOrderState> {
                         </div>
                     </div>
                 </div>
-                <div className="clearfix mx-auto" style={{width: 355, height: 108}}>
+                <div className="clearfix mx-auto" style={{width: 315, height: 108}}>
                    <div className="col col-7" style={{maxWidth: 235}}>
                        <TokenAmountInput
                            label="Fill amount"
@@ -366,7 +333,7 @@ export class FillOrder extends React.Component<FillOrderProps, FillOrderState> {
     private onFillAmountChange(isValid: boolean, amount?: BigNumber.BigNumber) {
         this.props.dispatcher.updateOrderFillAmount(amount);
     }
-    private onFillOrderChanged(e: any) {
+    private onFillOrderJSONChanged(e: any) {
         const orderJSON = e.target.value;
         this.setState({
             didOrderValidationRun: _.isEmpty(orderJSON) && _.isEmpty(this.state.orderJSONErrMsg),
